pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixeltunes64"
version = "0.2.0"
description = "Display Spotify cover art on a Raspberry Pi RGB matrix"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0.0",
    "spotipy>=2.25.0",
    "rgbmatrix @ git+https://github.com/hzeller/rpi-rgb-led-matrix.git",
]
